PR: Develop to master (#4)
* Added: Workflows badges

* Feature/get pages (#1)

* Added: Getting all pages from search

* Added: asks user

Now the application'll ask the user by the search term. Also, the handler for the response now is in the response_handler file

* Fixed: Docker build

* Refactored: Code structure (#2)

* Added: main content extractor

Now the program can extract the main content from the selected source. Its important to note that each website has your own frontend, meaning any css_selector can be used globally.

* Commented: The config module

It'll be used on future feature

* Added: CLI tool (#3)

* Added: CLI tool

Now the program is a CLI tools supporting flags and options. The dockerfile were updated with an entrypoint to support the arguments. The documentation were updated as well, side-by-side with the documentation on --help flag

* Updated: build Workflow name

* Bugfix/master workflows (#5)

* Added: CLI tool

Now the program is a CLI tools supporting flags and options. The dockerfile were updated with an entrypoint to support the arguments. The documentation were updated as well, side-by-side with the documentation on --help flag

* Updated: build Workflow name

* Fixed: Master with no workflow